Makefile fish support & fixes

+++ Makefile
@@ -6,13 +6,12 @@ default:
 
 setup:
 	mkdir -p ~/.config
-	ln -s helix ~/.config/helix
-	ln -s neofetch ~/.config/neofetch
-	ln -s atuin ~/.config/atuin
-	ln -s zsh/.zshrc ~/.zshrc
-	ln -s zsh/.zprofile ~/.zprofile
+	[ ! -e ~/.config/helix ] && ln -s helix ~/.config/helix || true
+	[ ! -e ~/.config/neofetch ] && ln -s neofetch ~/.config/neofetch || true
+	[ ! -e ~/.config/atuin ] && ln -s atuin ~/.config/atuin || true
+	[ ! -e ~/.config/fish ] && ln -s fish ~/.config/fish || true
 
-PACKAGES = helix neofetch zsh atuin autojump
+PACKAGES = helix neofetch zsh atuin autojump fish
 
 install-ubuntu: $(foreach pkg,$(PACKAGES),install-$(pkg)-ubuntu)
 
@@ -26,23 +25,29 @@ install-neofetch-ubuntu:
 	echo "Installing neofetch"
 	apt install neofetch
 
-install-zsh-ubuntu:
-	echo "Installing zsh"
-	apt install zsh
-	echo "Installing ohmyzsh"
-	sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
-	echo "Installing p10k"
-	git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+# install-zsh-ubuntu:
+# 	echo "Installing zsh"
+# 	apt install zsh
+# 	echo "Installing ohmyzsh"
+# 	sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+# 	echo "Installing p10k"
+# 	git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
 
 install-atuin-ubuntu:
+	echo "Installing atuin"
 	cargo install atuin
 
 install-autojump-ubuntu:
+	echo "Installing autojump"
 	mkdir -p build
 	git clone https://github.com/wting/autojump.git build/autojump
 	pushd build/autojump
 	./install.py
 	popd
+
+install-fish-ubuntu:
+	echo "Installing fish"
+	apt install fish
 
 clean:
 	rm ~/.config/helix
@@ -50,3 +55,4 @@ clean:
 	rm ~/.config/atuin
 	rm ~/.zshrc
 	rm ~/.zprofile
+	rm -r build
